# docs/swagger.yaml
basePath: /
definitions:
...
  current_event.CurrentEventDTO:
    properties:
      planItem:
        $ref: '#/definitions/current_event.PlanItemDTO'
      startTime:
        type: string
    type: object
  current_event.PlanItemDTO:
    properties:
      budgetItemId:
        type: integer
      name:
        type: string
      weeklyDuration:
        type: integer
    type: object
  rest.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  stats.DailyStatsDTO:
    properties:
      date:
        type: string
      perPlanItem:
        items:
          $ref: '#/definitions/stats.PlanItemStatsDTO'
        type: array
      totalTime:
        type: integer
    type: object
  stats.PlanItemStatsDTO:
    properties:
      duration:
        type: integer
      remaining:
        type: integer
      weeklyPlanItem:
        $ref: '#/definitions/weekly_plan.WeeklyPlanItemDTO'
    type: object
  stats.StatsSummaryDTO:
    properties:
      endDate:
        type: string
      perDay:
        items:
          $ref: '#/definitions/stats.DailyStatsDTO'
        type: array
      perPlanItem:
        items:
          $ref: '#/definitions/stats.PlanItemStatsDTO'
        type: array
      startDate:
        type: string
      totalPlanned:
        type: integer
      totalRemaining:
        type: integer
      totalTime:
        type: integer
    type: object
...
  weekly_plan.WeeklyPlanItemDTO:
    properties:
      budgetItemId:
        type: integer
      color:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      position:
        type: integer
      weeklyDuration:
        type: integer
      weeklyOccurrences:
        type: integer
    type: object
host: localhost:8181
info:
  contact: {}
  description: Time planning and tracking API
  title: Klokku API
  version: "0.9"
paths:
...
  /api/event:
    post:
      consumes:
      - application/json
      description: Start a new event with the given budget item id
      parameters:
      - description: Budget item id
        in: path
        name: budgetItemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/current_event.CurrentEventDTO'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Start a new event
      tags:
      - CurrentEvent
  /api/event/current:
    get:
      description: Retrieve the currently active event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/current_event.CurrentEventDTO'
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: No current event
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get the current running event
      tags:
      - CurrentEvent
  /api/event/current/start:
    patch:
      consumes:
      - application/json
      description: Update the start time of the currently running event
      parameters:
      - description: Start time in RFC3339 format
        in: body
        name: startTime
        required: true
        schema:
          properties:
            startTime:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/current_event.CurrentEventDTO'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: No current event
          schema:
            type: string
      security:
      - XUserId: []
      summary: Modify current event start time
      tags:
      - CurrentEvent
...
  /api/stats/weekly:
    get:
      description: Retrieve statistics for a specific week including time spent per
        plan item
      parameters:
      - description: Date in RFC3339 format (can be any day of the week)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/stats.StatsSummaryDTO'
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get weekly statistics
      tags:
      - Stats
...
securityDefinitions:
  XUserId:
    description: User ID header required for authentication
    in: header
    name: X-User-Id
    type: apiKey
swagger: "2.0"